+++ pochidetection/scripts/ssdlite/infer.py
@@ -163,6 +163,7 @@ def _setup_pipeline(
 
     transform = v2.Compose(
         [
+            v2.Resize(image_size),
             v2.ToImage(),
             v2.ToDtype(torch.float32, scale=True),
         ]